Exception Handling fix

+++ src/main/java/com/educandoweb/course/services/ProductService.java
@@ -1,59 +1,77 @@
 package com.educandoweb.course.services;
 
+import jakarta.persistence.EntityNotFoundException;
+
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.educandoweb.course.dto.ProductDTO;
 import com.educandoweb.course.entities.Product;
 import com.educandoweb.course.repositories.ProductRepository;
+import com.educandoweb.course.services.exceptions.DatabaseException;
+import com.educandoweb.course.services.exceptions.ResourceNotFoundException;
 
 @Service
 public class ProductService {
-    
+
     @Autowired
     private ProductRepository repository;
-    
+
     //Trás uma lista paginada dos produtos
     @Transactional(readOnly = true)
-    public Page<ProductDTO> findAll(Pageable pageable){
+    public Page<ProductDTO> findAll(Pageable pageable) {
         Page<Product> result = repository.findAll(pageable);
         return result.map(x -> new ProductDTO(x));
     }
-    
+
     //Trás um produto específico pelo seu ID
     @Transactional(readOnly = true)
-    public ProductDTO findById(Long id){
-        Product product = repository.findById(id).get();        
+    public ProductDTO findById(Long id) {
+        Product product = repository.findById(id).orElseThrow(
+                () -> new ResourceNotFoundException("ID não encontrado."));
         return new ProductDTO(product);
     }
-    
+
     //Recebendo DTO do produto cadastrado pelo frontend, salvando no banco de dados, e retornando a entidade produto convertido em DTO
     @Transactional()
-    public ProductDTO insert(ProductDTO dto){
-        Product entity = new Product();        
-        copyDtoToEntity(dto, entity);        
-        entity = repository.save(entity);
-        return new ProductDTO(entity);
-    }
-    
-    @Transactional()
-    public ProductDTO update(Long id,ProductDTO dto){
-        Product entity = repository.getReferenceById(id);
+    public ProductDTO insert(ProductDTO dto) {
+        Product entity = new Product();
         copyDtoToEntity(dto, entity);
         entity = repository.save(entity);
         return new ProductDTO(entity);
     }
-    
-    //Trás um produto específico pelo seu ID
+
     @Transactional()
-    public void delete(Long id){
-        repository.deleteById(id);
+    public ProductDTO update(Long id, ProductDTO dto) {
+        try {
+            Product entity = repository.getReferenceById(id);
+            copyDtoToEntity(dto, entity);
+            entity = repository.save(entity);
+            return new ProductDTO(entity);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("ID "+ id +" não encontrado. Não será possível atualizar os dados.");
+        }
     }
 
-    private void copyDtoToEntity(ProductDTO dto, Product entity) {                
+    //Trás um produto específico pelo seu ID
+    @Transactional(propagation = Propagation.SUPPORTS)
+    public void delete(Long id) {
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("ID solicitado não existe. Não será possível excluir.");
+        }
+        try {
+            repository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
+    }
+
+    private void copyDtoToEntity(ProductDTO dto, Product entity) {
         entity.setName(dto.getName());
         entity.setDescription(dto.getDescription());
         entity.setPrice(dto.getPrice());
